samples/agent/adk/quiz_tutor/pyproject.toml
[project]
name = "a2ui-quiz-tutor"
version = "0.1.0"
description = "Sample ADK-based Quiz Tutor agent that generates A2UI UI and is hosted as an A2A server agent."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "jsonschema>=4.0.0",
    "a2ui",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

